PLATFORM-4900: Allow to use only with access token and no id-secret
